pyproject.toml
[tool.poetry]
name = "ml4keiba"
version = "0.0.1"
description = "Machine Learning for 競馬 Dataset"
authors = ["Ningensei848 <k.kubokawa@klis.tsukuba.ac.jp>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
lxml = "^4.8.0"
beautifulsoup4 = "^4.10.0"
pandas = "^1.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
